+++ src/api/server.js
@@ -11,13 +11,12 @@ import gw2Api from 'lib/gw2';
 import tokenControllerFactory from './controllers/gw2-token';
 import usersControllerFactory from './controllers/user';
 import characterControllerFactory from './controllers/character';
+import pvpControllerFactory from './controllers/pvp';
 
 const authControllerFactory = require('./controllers/auth');
 const sitemapControllerFactory = require('./controllers/sitemap');
 const statisticsControllerFactory = require('./controllers/statistics');
-
 const CheckController = require('./controllers/check');
-const PvpController = require('./controllers/pvp');
 
 export default function createServer (models: Models, config: any) {
   createValidator.addDefaultRules();
@@ -58,7 +57,6 @@ export default function createServer (models: Models, config: any) {
   const gw2Tokens = tokenControllerFactory(models, createValidator, gw2Api);
 
   const checks = new CheckController(createValidator);
-  const pvp = new PvpController(models, gw2Api);
 
   const server = restify.createServer({
     name: 'api.gw2armory.com',
@@ -85,13 +83,13 @@ export default function createServer (models: Models, config: any) {
   });
 
   require('./resources')(server);
-  require('./resources/pvp')(server, pvp);
   require('./resources/guilds')(server, models);
   require('./resources/search')(server, models);
   require('./resources/users/check')(server, checks);
   require('./resources/users/gw2-token')(server, gw2Tokens);
   require('./resources/sign-upload')(server, models);
 
+  require('./resources/pvp')(server, pvpControllerFactory(models));
   require('./resources/characters')(server, characterControllerFactory(models));
   require('./resources/statistics')(server, statisticsControllerFactory(models));
   require('./resources/users').default(server, usersControllerFactory(models));