+++ src/config/default.js
@@ -55,5 +55,6 @@ module.exports = {
   cache: {
     findAllCharacters: 1,
     statistics: 1,
+    readLatestPvpSeason: 60000 * 60 * 1, // 1 hour
   },
 };