+++ src/fetch/fetchers/pvpStandings.js
@@ -8,17 +8,22 @@ export default async function fetchPvpStandings (models: Models, { token }: Fetc
 
   await Promise.all(standings.map((standing) => {
     const row = {
+      apiToken: token,
       seasonId: standing.season_id,
+
       totalPointsCurrent: standing.current.total_points,
       divisionCurrent: standing.current.division,
       pointsCurrent: standing.current.points,
       repeatsCurrent: standing.current.repeats,
       ratingCurrent: standing.current.rating,
+      decayCurrent: standing.current.decay,
+
       totalPointsBest: standing.best.total_points,
       divisionBest: standing.best.division,
       pointsBest: standing.best.points,
       repeatsBest: standing.best.repeats,
-      apiToken: token,
+      ratingBest: standing.best.rating,
+      decayBest: standing.best.decay,
     };
 
     return models.PvpStandings.upsert(row, {